deepen: SessionView owns served+drift (inline served-store/drift)
Inline served-store (281, _mergeServedRows merge invariant) and drift
(166, computeDrift/scanDrift with read+write side effect) plus workspace
(AsyncLocalStorage) and dsh-context (execCwd/execSessionKey) into
session-view.ts (now 322 lines, deep). Previously split across 4 files
with the drift write hidden inside a notice module; now co-located — the
served rows + reported drift + position reconstruction invariant lives in
one seam. Deleting it would scatter the served+drift invariant (deep).

Shallow files served-store/drift/workspace/dsh-context become 5-line shims
re-exporting from session-view for backward compat. No circular import
(session-view imports from hash-store/hash-assign, not from old shims).
Typecheck + 674 tests pass.
